Refactor cart and chat components: replace CartList with CartItem, update Chatv2 to Chat, and remove unused files for improved clarity and organization.
